[Telink] Add persistent subscriptions by default (#42452)
* [Telink] set persistent subscriptions by default

Now included to all Telink targets.

Persists Matter subscriptions on the publisher node. This feature allows
a Matter node to faster get back to the previous operation after it went
offline, for example, due to a power outage. That is, the node can use the
persisted subscription information to quickly re-establish the previous
subscriptions instead of waiting for the subscriber node to realize that
the publisher is alive again.

* [Telink] added dependenant configs & Buteo condition

Exclude persistent subscription for Buteo with retention
due to footprint.
Added dependencies for CHIPPlatformConfig.h

* Restyled by whitespace

---------

Co-authored-by: Restyled.io <commits@restyled.io>