Added two new Linux platform gn build arguments.
This adds two new Linux-specific gn build arguments that, in the
fullness of time, will influence how the Connectivity Manager is built
for that platform:

    1. chip_linux_network_manager
        * The Linux high-level network manager, if any, that owns network
          policy and state and orchestrates connectivity decisions and
          cross-interface policy.
        * At present, this must be "wpa_supplicant".

    2. chip_linux_wifi_paf_manager
        * The Linux Wi-Fi Neighbor Awareness Networking (NAN)
          Unsynchronized Service Discovery (USD) / Public Action Frame
          (PAF) commissioning transport manager.
        * At present, this must be empty or "wpa_supplicant".

Matter on Linux currently supports as first-class network
technologies: Bluetooth, Thread, Wi-Fi, and Ethernet with Bluetooth
reserved only for commissioning, and Thread, Wi-Fi, and Ethernet as
either commissioning transports, on-network commissioning networks,
and/or operational networks.

On Linux, BlueZ and the bluetoothd daemon is the de facto Bluetooth
implementation.

For Thread, OpenThread is the de facto Thread implementation for all
of Matter.

However, for Ethernet and Wi-Fi, things are more diverse on Linux.

In the nominal case (and currently in Matter), the Ethernet control
plane has no management software whatsoever beyond the Linux kernel
and the Wi-Fi control plane is handled by the wpa_supplicant daemon
over D-Bus.

However, there are several further variants:

    1. Instead of wpa_supplicant, Wi-Fi may be managed by iwd.
    2. Rather than being managed heterogeneously, as they are today in
       Matter, Ethernet, Wi-Fi, and, in some instances Cellular, may
       be further managed by one of:
       a. connman with:
          i.   Ethernet sub-managed by connman itself.
          ii.  Wi-Fi sub-managed by either:
               A. iwd
               B. wpa_supplicant
          iii. Cellular managed by oFono.
       b. NetworkManager with:
          i.   Ethernet sub-managed by NetworkManager itself.
          ii.  Wi-Fi sub-managed by either:
               A. iwd
               B. wpa_supplicant
          iii. Cellular managed by ModemManager.
    3. All of networking is managed by some proprietary product- or
       platform-specific entity for which any of the above may be a
       backing implementation.
       a. Regardless, like ConfigurationManager today, there may be a
          desire to default the Matter build to (1) or (2) but then set
          a custom ConnectivityManager at runtime with a potential
          chip::DeviceLayer::SetConnectivityMgr.